python-batch/python-scraping/requirements.txt
PyVirtualDisplay==0.2.1
selenium==3.4.3
Pillow==5.3.0
cx_Oracle==7.0.0
